+++ src/tests/js_denotest/Test_dbfs.js
@@ -11,7 +11,7 @@
 
 import { assertThrows, assertAlmostEquals } from "jsr:@std/assert@1";
 
-import { ONE_FRAME_SAMPLES } from "../../static/packet_conv/AUDIO_PARAM.js";
+import { ONE_FRAME_SAMPLES } from "../../static/AUDIO_PARAM.js";
 import { dbfs_float } from "../../static/dbfs.js";
 import { generate_rand_float32array } from "./util/rand_f32a.js";
