+++ src/tests/js_denotest/packet_conv/Test_audio.js
@@ -11,7 +11,7 @@
 
 import { assertEquals, assertThrows } from "jsr:@std/assert@1";
 
-import { ONE_FRAME_SAMPLES } from "../../../static/packet_conv/AUDIO_PARAM.js";
+import { ONE_FRAME_SAMPLES } from "../../../static/AUDIO_PARAM.js";
 import { packet_audio_encode, packet_audio_decode, is_audio_packet, AUDIO_PACKET_TYPE_ID, SILENT_AUDIO_PACKET_TYPE_ID } from "../../../static/packet_conv/audio.js";
 import { generate_rand_float32array } from "../util/rand_f32a.js";
 import { assertAlmostEqualsArray } from "../util/assertAlmostEqualsArray.js";
